Added dialogue reactions and greetings for the animals
Added a state machine for the case file so that it can slide up and down.
Changed animal to be on the UI rather than with sprites so that the dialogue would be on the same layer as the images.
Made everything in the scene into a prefab.
Decision buttons only show when the player can make a decision.